% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilters.R
\name{applyFilter}
\alias{applyFilter}
\title{Constructing robust reference standards for Nof1 studies for precision medicine}
\usage{
applyFilter(gs.mat = utils::data("mcf7"), expressionCutoff =30,
                   fold_change_window=c(1, 1.5), FDR.cutoff=0.1)
}
\arguments{
\item{gs.mat}{a matrix of paired samples used to develop the reference standard.
All paired samples must be adjacent}

\item{expressionCutoff}{a scalar positive integer > 1 indicating the minimum count value of RNA-seq}

\item{fold_change_window}{a tuple indicating the fold change range to search}

\item{FDR.cutoff}{fdr.threshold for determining which set of genes are DEGs in the reference standard}
}
\description{
\code{referenceNof1} is the R implementation of the reference biomarker algorithm by (Zaim 2020)
}
