% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_heatmap.R
\name{concordance_heatmap}
\alias{concordance_heatmap}
\title{Constructing robust reference standards for Nof1 studies for precision medicine}
\usage{
concordance_heatmap(jaccard_matrix)
}
\arguments{
\item{jaccard_matrix}{the concordance matrix used to create the heatmap}
}
\description{
\code{referenceNof1} is the R implementation of the reference biomarker algorithm by (Zaim 2020)
}
