% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.R
\name{fc}
\alias{fc}
\title{random forest feature selection based on binomial exact test}
\usage{
fc(pair.mat)
}
\arguments{
\item{pair.mat}{a two-column, stimulus-response paired sample matrix}
}
\value{
a data.frame with 1 columns containing the fold change
}
\description{
\code{referenceNof1} is the R implementation of the reference biomarker algorithm by (Zaim 2020)
}
