% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_similarity.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{random forest feature selection based on binomial exact test}
\usage{
jaccard(a,b)
}
\arguments{
\item{a}{a vector of DEG calls}

\item{b}{a vector of DEG calls}
}
\value{
a scalar indicating the jaccard similarity
}
\description{
\code{referenceNof1} is the R implementation of the reference biomarker algorithm by (Zaim 2020)
}
