% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_fc.R
\name{log_fc}
\alias{log_fc}
\title{random forest feature selection based on binomial exact test}
\usage{
log_fc(pair.mat)
}
\arguments{
\item{pair.mat}{a two-column, stimulus-response paired sample matrix}
}
\value{
a data.frame with 1 columns containing the log fold change
}
\description{
\code{referenceNof1} is the R implementation of the reference biomarker algorithm by (Zaim 2020)
}
