% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf7_data.R
\docType{data}
\name{mcf7}
\alias{mcf7}
\title{MCF7 Breast Cancer cell line}
\format{
An object of class \code{"data.frame"};
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE51403}{NCBI Repository}
}
\usage{
data(mcf7)
}
\description{
MCF7 Breast Cancer cell line experiment that examines expression profiling by high throughput sequencing to determine whether
more sequence, or more replication is required to obtain more accurate results for differential gene expresion. Contains the first
the first 5,000 transcripts as an example of the dataset.
}
\details{
Experimental design: Treatment (10nM E2 treatment for 24h) and control MCF7 cells are both replicated 7 times,
and collected for mRNA-seq. Reads are then subsampled for statistical analysis.
}
\examples{
data(mcf7)
colMeans(mcf7)
summary(mcf7)
}
\references{
Liu Y, Zhou J, White KP. RNA-seq differential expression studies: more sequence or more replication? Bioinformatics 2014 Feb 1;30(3):301-4. PMID: 24319002
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
