% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameters.R
\name{optimize_parameters}
\alias{optimize_parameters}
\title{Constructing robust reference standards for Nof1 studies for precision medicine}
\usage{
optimize_parameters(gs.mat=utils::data("mcf7"),
                           expressionCutoffs = c(30,50),
                           FCRegions = list(c(0,0.3)),
                           FDR.cutoff = 0.1)
}
\arguments{
\item{gs.mat}{the matrix used to develop the reference standard}

\item{expressionCutoffs}{a vector of positive integer > 1 indicating the minimum count value of RNA-seq}

\item{FCRegions}{a tlist of tuples indicating the fold change region to search}

\item{FDR.cutoff}{fdr.threshold for determining which set of genes are DEGs in the reference standard}
}
\description{
\code{referenceNof1} is the R implementation of the reference biomarker algorithm by (Zaim 2020)
}
